+++ client/src/scripts/others/worker_server.ts
@@ -1,6 +1,6 @@
 import { BattleRoyale, LevelPlayer, OfflineGameServer } from "./offline.ts";
 import { GameServerConfig } from "common/scripts/config/config.ts";
-import { FetchFileManager, OfflineClientsManager, WorkerSocket } from "common/engine/core.ts";
+import { FetchFileManager, OfflineClientsManager, Path, WorkerSocket } from "common/engine/core.ts";
 import { PacketManager } from "common/scripts/packets/packet_manager.ts";
 let server:OfflineGameServer
 let level:LevelPlayer
@@ -18,8 +18,18 @@ self.onunhandledrejection = (ev) => {
     logError(ev.reason)
 }
 
+
+async function load_level(path:string){
+    const fs=new FetchFileManager()
+    fs.base=path
+    level=new LevelPlayer(server,fs)
+    await level.begin(path)
+    self.postMessage({
+        type:"server_created"
+    })
+}
 self.onmessage = async(ev) => {
-    const msg = ev.data;
+    const msg = ev.data
     switch(msg.type){
         case "begin":{
             server=new OfflineGameServer(
@@ -29,14 +39,7 @@ self.onmessage = async(ev) => {
             break
         }
         case "load_level":{
-            const path="/"+msg.path
-            const fs=new FetchFileManager()
-            fs.base=path+"/"
-            level=new LevelPlayer(server,fs)
-            await level.begin(path)
-            self.postMessage({
-                type:"server_created"
-            })
+            load_level(msg.path)
             break
         }
         case "load_mode":{
@@ -73,6 +76,15 @@ self.onmessage = async(ev) => {
         case "connect":{
             const ws=new WorkerSocket(self as unknown as Worker)
             server.clients.fake_connect_other_s(ws)
+            break
+        }
+        case "next_level":{
+            if(!level)break
+            self.postMessage({
+                type:"start_level",
+                path:Path.join(level.path,(level.level.next_level as Record<string,string>)[msg.name]),
+                start_with_intro:msg.start_with_intro
+            })
             break
         }
     }