+++ common/scripts/definitions/loot_tables.ts
@@ -73,16 +73,16 @@ export const LootTables={
         {table:"death_consumibles",count:1,weight:0.04},
     ],
     health_consumibles:[
-        {item:"bandage",count:5,weight:6},
+        {item:"bandage",count:5,weight:5},
         {item:"medikit",weight:1},
     ],
     adrenaline_consumibles:[
-        {item:"yellow_soda",count:1,weight:8},
+        {item:"yellow_soda",count:1,weight:6},
         {item:"inhaler",count:1,weight:1},
         {item:"yellow_pills",count:1,weight:0.1},
     ],
     shield_consumibles:[
-        {item:"blue_soda",count:2,weight:8},
+        {item:"blue_soda",count:2,weight:6},
         {item:"blue_potion",count:1,weight:1},
         {item:"blue_pills",count:1,weight:0.1},
     ],
@@ -92,7 +92,7 @@ export const LootTables={
         {item:"red_crystal",count:2,weight:0.1},
     ],
     mana_consumibles:[
-        {item:"purple_soda",count:2,weight:8},
+        {item:"purple_soda",count:2,weight:6},
         {item:"purple_potion",count:1,weight:1},
         {item:"purple_pills",count:1,weight:0.1},
     ],