+++ common/scripts/definitions/maps/tundra.ts
@@ -478,7 +478,7 @@ export const TundraMap:MapDef={
 
             {def:"tundra_crate",count:7},
             {def:"copper_crate",count:20},
-            {def:map_spawns.crates,count:80},
+            {def:map_spawns.crates,count:100},
             {def:[
                 {def:"oak_tree",weight:100},
                 {def:"pine_tree",weight:75},